+++ changelog.md
@@ -7,6 +7,7 @@
 - Fix for music showing all "concluded"
 - Fix for incorrect time used for "now"
 - More changed to Home
+- Fix listing historical events in favorites and music
 
 ## 1.38
 - Change to use standalone components